mononoke: update walker loaded count log ready for chunking
Summary:
Calling the count Final doesn't make sense when chunking as it can appear more than once per tail, so update it to something more appropriate.

Seen is the count of Nodes seen in the stream. Loaded is the count of nodes where NodeData was loaded.  When chunking some deferred Nodes will be seen but not Loaded.

Reviewed By: krallin

Differential Revision: D25742283

fbshipit-source-id: 1f10007d94ad2dbd750bfa53bab3e46a2caad7fa