mononoke: fix walker stats delta per s in last log of run
Summary:
Include changes since the last progress update (if any) in the final delta and total time logged.

When not chunking this was a minor inaccuracy, but when chunking in small chunks it would mean a lot of stats for each chunk were missing as if a chunk took less than 5 seconds there might not even have been a progress update before this end state.

Reviewed By: StanislavGlebik

Differential Revision: D25852274

fbshipit-source-id: 6cea76e3abd37908475052947794eed442a1ac82