eden: log stack traces in debugimporthelper
Summary:
I've been seeing this hundreds of times / day in my Eden logs:

```
2021-01-07 09:19:19,702 Fetching failed, continuing as this may be spurious: Expected type that converts to PySequence but received set
```

Unfortunately, those logs are really not very helpful, because that error
doesn't tell you anything about where it occurred. I do however care about them
because they have correlated with slowness in Eden for me :)

At the same time, this error is a bit important: The last time we had an issue
here, it turned into a SEV. That was because the fallback was buggy, but it
does suggest that we do not expect to be routinely falling back.

So, let's log more.

Reviewed By: xavierd

Differential Revision: D25924077

fbshipit-source-id: f59d81cc51786f94f747142ec4baaca78f3072c2