configparser: add support for bypassing all shards
Summary:
quark-zju made it much safer to rollout configs recently by enforcing
time-sharding, but one little downside is that it becomes harder to test
them locally (or at diff-time, which I'm working on adding) because the time
shard will usually not pass.

This diffs add an environment variable to force all shards to pass.

Reviewed By: DurhamG

Differential Revision: D25925207

fbshipit-source-id: 343b90165ad2b6bae64b6821eae95c58f7d79698