mononoke: disallow pushing public commits with change-xrepo-mapping extra
Summary:
In the previous diff we added a special commit extra which allows backsyncer to
change the mapping version that was used to rewrite commits between repos.

This is a dangerous thing since it might corrupt the repository. So to avoid
people accidentally corrupting the repository let's add a check that would
prevent landing public commits with change-xrepo-mapping extra.

Reviewed By: ahornby

Differential Revision: D25951529

fbshipit-source-id: d2de37d1bf2f2eb9f45b3bf3e21706a799a50f69