mononoke: make it possible to change mapping version via commit extra
Summary:
At the moment the procedure to change commit sync mapping version is fairly
complicated [1]. This diff attempts to simplify one aspect of it.

At the moment we need to disable pushredirection and manually sync a commit
between two repos. Disabling pushredirection is quite error prone and led to
SEVs in the past. With this diff we just need to create a commit with a special
extra "change-xrepo-mapping" and let the backsyncer rewrite a commit with the
new version for us.

This means that landing a commit with this extra should not be allowed to
everyone - I'm going to change bookmark movement to check that.

Reviewed By: ikostia

Differential Revision: D25950531

fbshipit-source-id: 78b1f4950e6bb7da8b511e02685ed6084c29a680