http persistent connection for netspeedtest requests
Summary:
This is implements http keep-alive behavior for `/netspeedtest` endpoint. `keep-alive` is implicit for `HTTP/1.1`. It means that connection is not closed after sever sends a response. Further request/response pairs can be exchanged.

This enables warmup measurements in netspeedtest before real measurements are made.

The difference is quite significant, even when I am measuring this server-side.

Reviewed By: krallin

Differential Revision: D25946675

fbshipit-source-id: a7b0f812bf61d704ca93615ab09d0897a08f1ba1