mononoke/hgcli: shut down even harder
Summary:
Updating to a newer version of Tokio breaks hgcli.

This seems to be happening because Tokio fixed a bug where shutting down your
runtime could leak threads (https://github.com/tokio-rs/tokio/pull/2649), which
results in the runtime refusing to shutdown if you have a thread waiting to
read on stdin (note: I didn't try to back that out specifically to confirm).

Unfortunately, we have that in hgcli, because we usually kick off a stdin read
before we know if we'll need the data or not.

Here's the backtrace for a blocked shutdown P163996251.

We  were already trying to shut down the runtime and ignore every outstanding
work (because we ensure earlier that there is nothing outstanding), so let's
just try even harder.

Reviewed By: StanislavGlebik

Differential Revision: D25952430

fbshipit-source-id: ae3a1413790cf81a5d990220f2dc5d4599219f73