mononoke: remove BlobConfig::Scrub variant
Summary:
Scrub was being enabled by mutating the BlobConfig into the BlobConfig::Scrub.

This change removes the BlobConfig::Scrub variant as its not present in the thrift config any more, so there is not need for it to exist. Scrubbing is an optional mode of multiplexed store construction, so instead add ScrubOptions to BlobstoreOptions.

Reviewed By: StanislavGlebik

Differential Revision: D25927253

fbshipit-source-id: 29fceae59be8f7068300a7b8b298c6edbe4da04b