mononoke: add a library for ALPN support
Summary:
I'd like to update our connection acceptor to support ALPN in order to
distinguish hgcli traffic from http traffic (the former will have ALPN and the
latter will not).

To do so, I need to parse ALPN, which is a length-prefixed bytestring format.
Let's add support for it in a library. This is what this does.

Reviewed By: johansglock

Differential Revision: D25954536

fbshipit-source-id: 3756dde90906a5f0f36630b866bcb17d03bc94c6