mononoke: walk hg filenodes in chunks
Summary:
The test added in previous diff showed that hg filenodes weren't being deferred between chunks in the expected way.

This is because we can't tell if a hg filenode is in a given chunk until it is loaded. This is similar to unodes, but the linked changeset in this case isHgChangesetId rather than the bonsai ChangesetId, so this change introduces hg_to_bcs mapping in the walker state, which is used for looking up whether the filenodes linked HgChangesetId is in the chunk, and if not defers the edge.

Reviewed By: krallin

Differential Revision: D25742276

fbshipit-source-id: 1f92452d012aab5b9fdf29f43fc05ebc043b2c7a