mononoke: include root count in walker queue stats
Summary:
The main thing this change does is make sure pending roots to visit are represented in the difference between Walked and Children.  Children is the sum of all child nodes discovered, both visited and unvisited.  Walked is a measure of number of nodes visited.  Children-Walked is used as a measure of queue depth remaining to be processed.

When not chunking this is a minor issue as usually just one bookmark root node is not counted in children, but when chunking not counting the roots means mean the chunk of several 100000 roots is not visible as waiting to be processed.

Reviewed By: StanislavGlebik

Differential Revision: D25852526

fbshipit-source-id: df5f21a37be152f0baee40d33fd7dfb7aaa763de