Add fields to Cargo.toml's to make crates publishable
Summary:
This diff adds some basic information. Note I have no idea what most of
these crates do so the descriptions are a bit light. I'd appreciate it if people
could tell me what the descriptions should be. Or better yet, just put up a diff
for it after this one goes in.

Reviewed By: lukaspiatkowski

Differential Revision: D26031870

fbshipit-source-id: 496f44882b6ed80a4f3834cf0fa3215ab318c930