adding hidden --preserve-mount-point flag to eden rm
Summary:
As described in D26033704, it would be really nice if we could have a version of `eden rm` that did everything except remove the mount point directory.  This creates a nice symmetry between the tolerated start and end points of a clone/rm cycle, in that `eden clone` already tolerates being targeted against and empty directory.

NOTE: I made no effort to make this work on Windows.  The implementation was different enough that I wasn't exactly sure how you all wanted to go about it.  We don't run on Windows at all, so this is all we need, but I'd understand if you all want to make it work, or at least throw an explanatory error.

Reviewed By: chadaustin

Differential Revision: D26033912

fbshipit-source-id: 36d63ad510393d3d09b1912a73efaebede534f61