mononoke: add ability to clear walker state
Summary:
When running large repos its interesting to be able to reduce memory usage by clearing all or part of the walk state between chunks.

This adds ability to clear the node data by node type, and to clear the interning maps by the type of interned data type

The only interned type not clearable is the bonsai changeset id as that is used to maintain the list of deferred edges for later chunks to process.

Reviewed By: krallin

Differential Revision: D25867960

fbshipit-source-id: 4b48f03b1a1b8fef1c5ded952bdcd6b1241dcc32