tracing: add edenscm.tracing Python module mirroring Rust APIs
Summary:
Add a `edenscm.tracing` Python module that mirrors some of the APIs in the Rust
tracing eco-system:
- span, {info,debug,trace,warn,error}_span (for tracing)
- event, {info,debug,trace,warn,error} (for logging)
- instrument (decorator for function tracing)

One thing that needs to be careful is that callsites are meant to be static:
multiple calls to a same code location should reuse a single callsite. That
is achieved using (id(code), line) from the Python stack as the key.

Reviewed By: sfilipco

Differential Revision: D26013750

fbshipit-source-id: 63b89a73186d876af9fc1b311fbd7948d0a6e27a