filesystem: reuse wctx/pctx during status tight loops
Summary:
When comparing file contents with the version on disk we were
recreating the wctx and parent ctx for each file. This prevents us from
benefiting from the caching inside the ctx object, and with the new changelog
backend (which lacks certain caching) it can slow down hg status by 5x when
dealing with 10k+ pending changes.

Let's reuse the ctx objects to avoid this.

Reviewed By: quark-zju

Differential Revision: D26135419

fbshipit-source-id: ac25c69c26965362d0e5bfa92288d21f3e230be7