mononoke/repo_listener: default to allowlist_checker if no ACL
Summary:
Right now, if we got no Hipster ACL in a repo listener, we default to denying
all access.

This is kind of annoying when using Mononoke locally but behind a trusted HTTP
proxy, because that means you cannot access the repo at all (if you're not
behind a proxy, then your TLS identities are used instead and everything is
fine).

If we trust a given proxy to impersonate literally anyone, we probably trust
them to access the repo to begin with (since the set of people with access is
not empty and therefore there is always at least someone they could impersonate
that has access), so this is what this diff does.

Reviewed By: johansglock

Differential Revision: D26073274

fbshipit-source-id: 0ef06cb6283d7f69072b712d3cb5a8383a493998