mononoke/repo_client: bring back `mod tests`
Summary:
This test module accidentally got lost when I added a `mod tests { ... }` in
the containing module. This brings it back and modernizes the tests that could
be. The push redirection test has way too much boilerplate to be manageable so
for now I removed it. I'll see if I can bring it back after some refactoring
I'm doing.

I'll try to see if there's a way we can try to lint / warn against inline
modules shadowing other files.

Reviewed By: ahornby

Differential Revision: D26124354

fbshipit-source-id: 7b24c4fe635bf8197142ab9ee087631ed49f10be