mononoke/mononoke_api: make WBC configurable + take fewer parameters
Summary:
Right now, if you enable Mononoke API, you always get a WBC, for all derived
data kinds, and with a delay. This isn't great for a few reasons:

- Some places don't really care about bookmarks being warm to being with. This
  is the case in the bookmarks filler (note that right now, it still does
  polling to satisfy the WBC API, it's just not "warm").
- Some places don't want a delay, or don't want all kinds. This is the case for
  Mononoke Server (which doesn't use Mononoke API right now, but that's what
  I'm working towards), or EdenAPI, which uses a WBC sort-of-de-facto but
  doesn't really care (but likely will in the future, and will want to follow
  Mononoke Server's behavior).

As of this diff, we can now configure this when initializing `Mononoke`. I also
split out all those args into a `MononokeEnvironment` because the args list
was getting out of hand. One thing I did not do is make a way to instantiate
`MononokeEnvironment` from args (but we should do it at some point in the
future).

Reviewed By: StanislavGlebik

Differential Revision: D26100706

fbshipit-source-id: 1daa6335f3ce2b297929a84788bc5b4d9ad6432f