mononoke: fix warm bookmark cache bug when no bookmark history is found
Summary:
It's possible that a bookmark has no history in its bookmark log. It shouldn't
happen normally, but it might happen in a few cases:
1) If bookmarks were imported before bookmark update log was added to mononoke
(unusual, but happens)
2) if bookmark update log was cleaned for some reason.

So if a bookmark wasn't warm at start, then single_bookmark_updater would never
warm it unless this bookmark gets new entry in boomkark update log. This diff
fixes it.

Reviewed By: krallin

Differential Revision: D26149435

fbshipit-source-id: 5bba8764050349adf106c0e68488981cf21055c4