nfs: unregister self with portmap on destruction
Summary:
When the RPC server starts, it registers itself against the global rpcbind
daemon to allow RPC clients to reach it. However, it never unregisters itself
when it is destructed, leading to the rpcbind daemon having outdated
information. It also prevents a future RPC server for the same program and
version to register itself.

To fix this, we can simply unregister on destruction. The MountdUtil binary is
also fixed to be able to interrupt it with ctrl+c.

Reviewed By: chadaustin

Differential Revision: D26105289

fbshipit-source-id: c9c3d552ec04a78025d6ff715a425f9ec3aa1328