nfs: plumbing for running an RPC server
Summary:
The NFS protocol is comprised of several different RPC "programs": the mount,
nlm and nfs programs. Since all of these 3 needs to independently register with
rpcbind, let's have some common scafolding to read/write from the socket and to
simplify the implementation of these programs.

This code was written by wez

Reviewed By: chadaustin

Differential Revision: D25986691

fbshipit-source-id: 15c5fdc68323fd964ed79aa06392a83bf964ab4a