mononoke/getbundle_response: log args to call_difference_of_union_of_ancestors_revset
Summary:
Like it says in the title, this logs arguments that we pass to
call_difference_of_union_of_ancestors_revset.

The underlying goal is to see if we could benefit from caching here by seeing
how unique the args are.

Facebook

Getbundle accounts for a small portion of our traffic, so Scuba-wise I think
this should be fine.

Reviewed By: StanislavGlebik

Differential Revision: D26202463

fbshipit-source-id: 93a82662764d0b114291d72ffc79d977c9721d63