bookmarks_filler: add option to skip connecting to scmquery
Summary:
Add an option to the bookmarks filler to skip checking SCMQuery.  Use this in
tests so that tests don't attempt to connect to SCMQuery.

Ensure the client id and type is filled in for the SCMQuery requests so that
these requests have appropriate values when requests do get sent to SCMQuery.

Reviewed By: mitrandir77

Differential Revision: D26222850

fbshipit-source-id: 620c30d1418f4e951687f244ec769b3e464462b4