mononoke: add walker checkpointing
Summary:
Walker checkpoints allow a scrub or other walk to continue where it left off.  This is useful as we can release new code without making the scrub start from scratch again.

This change adds checkpoint loading and recording to tail.rs along with a new test for it.

When restarting from a checkpoint the code considers the unfinished checkpoint itself as the main_bounds and and new commits since the checkpoint as the catchup_bounds.

If there is no checkpoint at all the repo bounds are used as the main_bounds.

Differential Revision: D25995106

fbshipit-source-id: e1663091e4b1157541b256f36b354bbf316a92c9