mononoke/server: move the scuba configuration to the CLI not the config
Summary:
We always log to the same destination, but this is a little annoying as it
stands for e.g. hooking in EdenAPI because that wants 1 Scuba dataset for the
whole thing instead of per-repo.

This also means that logging to Scuba for things that aren't tied to a repo
isn't possible, which might explain why our logging in connection_acceptor
is a) very sparse and b) only logs to stderr.

To make the transition smooth here, I added support for a default Scuba
table and used this in Mononoke Server. I think we can remove this after
updating our tasks to explicitly receive the dataset as an argument.

Reviewed By: StanislavGlebik

Differential Revision: D26126012

fbshipit-source-id: 0fa92dc8c5d5ddeed99dd7d9dd5a2288b8300bf3