scuba: add a way to log sampled out rows in verbose mode
Summary:
This allows us to log sampled messages, but reserves an option of falling back to full verbose logging in critical situations.

Note that while this might be a desired behavior in most cases, it's certainly not always the right thing to do: sometimes sampled data needs to remain sampled, even for verbose logging.

Reviewed By: ahornby

Differential Revision: D26148454

fbshipit-source-id: c6ff9d1b05c9cec4895181e008ef6483884bb483