mononoke: set minimum successful writes in integration tests
Summary:
Consider test-blobstore_healer.t. We write to 3 blobstores, and expect 1 to
fail all the time and the others to have all the blobs. Except this isn't
really the guarantee the multiplex makes when you configure it with just 1
store to write to (it guarantees the blob is in ONE store, not 2).

The relevant failure is: P172423216.

So let's have a config that reflects our expectations. There are other tests
that use the Multiplex. I'll let CI tell me if they need changing.

Hopefully this should eliminate one more source of test flakiness...

Reviewed By: StanislavGlebik

Differential Revision: D26275304

fbshipit-source-id: 9184420b2ce0990d91e4ec6d38d28ec5ae87775d