mononoke/server: bring back cache warmup parallelism (aka boot 2x as fast)
Summary:
Like it says in the title. This updates Mononoke to restore parallelism in the
repo startup sequence, notably as part as cache warmup is concerned, which is
the biggest time consumer of Mononoke startup (I'm saying "restore" here
because we used to have this and it appears to have been accidentally removed a
month ago).

Note that we incidentally got some of the parallelism back earlier this week
for repo creation when I updated Mononoke Server to use `mononoke_api`, but
that didn't do anything for cache warmup, hence this diff.

Reviewed By: StanislavGlebik

Differential Revision: D26276745

fbshipit-source-id: 687a52a2326de7158fc7ad1982547a698faff314